requirements.txt
markdownify==0.10.1
python-slugify==5.0.2